Proportions

The best Proportions package.

Package Features

Provides functions to provide proportionality for vector data

Documentation

Proportions.get_proportion_round_sq — Method

get_proportion_round_sq(for_prop_arr::Array{Float64, 1};round_digits::Int64)::Array{Float64, 1}

Compute the proportional values for each element in the array and round to a specified decimal digits. If the sum of all the elements is not 1 , then square the array and try looking for proportional correctness after 2 itterations , the adjustment is done in the 1st largest element returns a Array{BigFloat, 1}

Example

julia> get_proportion_round([1.33,1.33,1.23,1.1111],round_digits=2)
=[0.26, 0.27, 0.25, 0.22]